[#33] refactor : enums 도메인 분리
