+++ src/main/java/com/juu/juulabel/domain/enums/member/Gender.java
@@ -1,4 +1,4 @@
-package com.juu.juulabel.domain.enums;
+package com.juu.juulabel.domain.enums.member;
 
 public enum Gender {
     FEMALE, MALE, NONE
