+++ src/main/java/com/juu/juulabel/domain/enums/member/MemberRole.java
@@ -1,4 +1,4 @@
-package com.juu.juulabel.domain.enums;
+package com.juu.juulabel.domain.enums.member;
 
 public enum MemberRole {
     ROLE_ADMIN, ROLE_USER
