+++ src/main/java/com/juu/juulabel/domain/enums/member/MemberStatus.java
@@ -1,4 +1,4 @@
-package com.juu.juulabel.domain.enums;
+package com.juu.juulabel.domain.enums.member;
 
 import lombok.Getter;
