+++ src/main/java/com/juu/juulabel/domain/repository/jpa/MemberJpaRepository.java
@@ -2,7 +2,7 @@ package com.juu.juulabel.domain.repository.jpa;
 
 
 import com.juu.juulabel.domain.entity.member.Member;
-import com.juu.juulabel.domain.enums.Provider;
+import com.juu.juulabel.domain.enums.member.Provider;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
