+++ src/main/java/com/juu/juulabel/domain/repository/reader/MemberReader.java
@@ -4,7 +4,7 @@ import com.juu.juulabel.common.exception.InvalidParamException;
 import com.juu.juulabel.common.exception.code.ErrorCode;
 import com.juu.juulabel.domain.annotation.Reader;
 import com.juu.juulabel.domain.entity.member.Member;
-import com.juu.juulabel.domain.enums.Provider;
+import com.juu.juulabel.domain.enums.member.Provider;
 import com.juu.juulabel.domain.repository.jpa.MemberJpaRepository;
 import com.juu.juulabel.domain.repository.query.MemberQueryRepository;
 import lombok.RequiredArgsConstructor;
